move owner of zn_input_device
zn_input_device is owned by input_manager before, but now, it is owned by zn_seat.
This change means also removing zn_seat_device.
